requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
cffi==1.11.5
confusable-homoglyphs==3.1.1
decorator==4.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11
django-bootstrap3==10.0.1
django-braces==1.13.0
django-floppyforms==1.7.0
django-registration==2.4.1
django-tinymce==2.7.0
django-toolbelt==0.0.1
factory-boy==2.11.1
Faker==0.8.15
gunicorn==19.8.1
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
misaka==2.1.0
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.4
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
static3==0.7.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.1
